backend/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.5.0
async-timeout==4.0.3
attrs==24.2.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
exceptiongroup==1.2.2
Flask==3.0.3
Flask-Cors==5.0.0
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
multidict==6.1.0
openai==0.28.0
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
tqdm==4.66.5
twilio==9.3.1
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.0.4
yarl==1.11.1
zipp==3.20.2
